+++ src/current.rs
@@ -1,7 +1,7 @@
 use crate::assert_positive_float;
 use core::{cmp, fmt, ops};
 
-/// Represents a current value, stored as whole microamps (μA) stored in a `u64` value.
+/// Represents a current value, stored as whole microamps (μA) as a 64-bit value.
 /// This value can only be positive.
 ///
 /// **Reminder:** `1000 μA = 1 mA, 1000 mA = 1 A`
@@ -32,9 +32,9 @@ use core::{cmp, fmt, ops};
 /// let c2 = 1.2.amps(); // 1.2A
 ///
 /// if c1 > c2 {
-///     println!("{:?} is greater than {:?}", c1, c2);
+///     println!("{} is greater than {}", c1, c2);
 /// } else {
-///     println!("{:?} is less than or equal to {:?}", c1, c2);
+///     println!("{} is less than or equal to {}", c1, c2);
 /// }
 /// ```
 ///
@@ -47,15 +47,15 @@ use core::{cmp, fmt, ops};
 /// ```rust
 /// use ohms::prelude::*;
 ///
-/// let c1 = 500.amps(); // 0.5A
+/// let c1 = 500.milli_amps(); // 0.5A
 /// let c2 = 1.1.amps(); // 1.1A
 ///
-/// let c3 = c1 + c2; // 1.6A
-/// let c4 = c2 - 300.milli_amps(); // 0.8A
+/// let sum = c1 + c2; // 1.6A
+/// let diff = c2 - 300.milli_amps(); // 0.8A
 /// ```
 ///
 /// # Scaling Current values
-/// You can use the `*` and `/` operators to scale `Current` values by a scalar `u32` or `f32` value.
+/// You can use the `*` and `/` operators to scale `Current` values by an integer or floating-point value.
 /// The result is a new `Current` value, rounded down to the nearest whole microamp (μA).
 ///
 /// If the result of operation would overflow or underflow, the operation will panic.
@@ -66,7 +66,7 @@ use core::{cmp, fmt, ops};
 /// use ohms::prelude::*;
 ///
 /// let c1 = 200.milli_amps(); // 200mA
-/// let c2 = c1 * 3; // 660mA
+/// let c2 = c1 * 3; // 600mA
 ///
 /// let r3 = 1.5.amps(); // 1.5A
 /// let r4 = r3 / 2.5; // 0.6A
@@ -81,7 +81,7 @@ use core::{cmp, fmt, ops};
 ///
 /// let c1 = 1.2.amps(); // 1.2A
 ///
-/// println!("{:.3}A is {:.1}A", c1.amps(), c1.milli_amps());
+/// println!("{:.3} A is {:.1} mA", c1.amps(), c1.milli_amps());
 /// ```
 ///
 #[derive(Clone, Copy, Debug)]
@@ -95,8 +95,8 @@ impl Current {
     /// It is recommended to use the `micro_amps`, `milli_amps`, and `amps`, extension
     /// methods on integer and floating-point types instead.
     #[inline]
-    pub const fn from_micro_amps(value: u64) -> Current {
-        Current { raw: value }
+    pub const fn from_micro_amps(value: u64) -> Self {
+        Self { raw: value }
     }
 
     /// Returns the current value in whole microamps (μA).
@@ -126,13 +126,13 @@ impl Current {
     /// Returns a `Current` value of zero amps (0A).
     #[inline]
     pub const fn zero() -> Self {
-        Current::from_micro_amps(0)
+        Self::from_micro_amps(0)
     }
 }
 
 impl PartialEq for Current {
     #[inline]
-    fn eq(&self, other: &Current) -> bool {
+    fn eq(&self, other: &Self) -> bool {
         self.raw == other.raw
     }
 }
@@ -141,40 +141,40 @@ impl Eq for Current {}
 
 impl PartialOrd for Current {
     #[inline]
-    fn partial_cmp(&self, other: &Current) -> Option<cmp::Ordering> {
+    fn partial_cmp(&self, other: &Self) -> Option<cmp::Ordering> {
         self.raw.partial_cmp(&other.raw)
     }
 }
 
 impl Ord for Current {
     #[inline]
-    fn cmp(&self, other: &Current) -> cmp::Ordering {
+    fn cmp(&self, other: &Self) -> cmp::Ordering {
         self.raw.cmp(&other.raw)
     }
 }
 
 impl ops::Add for Current {
-    type Output = Current;
+    type Output = Self;
 
     /// Adds two `Current` values together, returning a new `Current` value.
     #[inline]
-    fn add(self, other: Current) -> Current {
+    fn add(self, other: Self) -> Self {
         self.raw
             .checked_add(other.raw)
-            .map(Current::from_micro_amps)
+            .map(Self::from_micro_amps)
             .expect("Overflow when adding current values")
     }
 }
 
 impl ops::Sub for Current {
-    type Output = Current;
+    type Output = Self;
 
     /// Subtracts one `Current` value from another, returning a new `Current` value.
     #[inline]
-    fn sub(self, other: Current) -> Current {
+    fn sub(self, other: Self) -> Self {
         self.raw
             .checked_sub(other.raw)
-            .map(Current::from_micro_amps)
+            .map(Self::from_micro_amps)
             .expect("Overflow when subtracting current values")
     }
 }
@@ -182,17 +182,17 @@ impl ops::Sub for Current {
 macro_rules! impl_mul_for_integer {
     ($i:ty) => {
         impl ops::Mul<$i> for Current {
-            type Output = Current;
+            type Output = Self;
 
             #[inline]
             #[allow(unused_comparisons)]
-            fn mul(self, scale_factor: $i) -> Current {
+            fn mul(self, scale_factor: $i) -> Self {
                 if scale_factor < 0 {
                     panic!("Cannot multiply current value by negative value")
                 }
                 self.raw
                     .checked_mul(scale_factor as u64)
-                    .map(Current::from_micro_amps)
+                    .map(Self::from_micro_amps)
                     .expect("Overflow when multiplying current value")
             }
         }
@@ -209,21 +209,21 @@ impl_mul_for_integer!(i32);
 impl_mul_for_integer!(i64);
 
 impl ops::Mul<f32> for Current {
-    type Output = Current;
+    type Output = Self;
 
     /// Multiplies a `Current` value by a floating-point value, returning a new `Current` value.
     #[inline]
-    fn mul(self, scale_factor: f32) -> Current {
+    fn mul(self, scale_factor: f32) -> Self {
         self * scale_factor as f64
     }
 }
 
 impl ops::Mul<f64> for Current {
-    type Output = Current;
+    type Output = Self;
 
     /// Multiplies a `Current` value by a floating-point value, returning a new `Current` value.
     #[inline]
-    fn mul(self, scale_factor: f64) -> Current {
+    fn mul(self, scale_factor: f64) -> Self {
         let result = match scale_factor {
             _ if scale_factor.is_infinite() => {
                 panic!("Cannot multiply current value by infinity")
@@ -235,19 +235,19 @@ impl ops::Mul<f64> for Current {
             _ => self.raw as f64 * scale_factor,
         };
 
-        Current::from_micro_amps(result as u64)
+        Self::from_micro_amps(result as u64)
     }
 }
 
 macro_rules! impl_div_for_integer {
     ($i:ty) => {
         impl ops::Div<$i> for Current {
-            type Output = Current;
+            type Output = Self;
 
             /// Divides a `Current` value by an integer value, returning a new `Current` value.
             #[inline]
             #[allow(unused_comparisons)]
-            fn div(self, divisor: $i) -> Current {
+            fn div(self, divisor: $i) -> Self {
                 if divisor == 0 {
                     panic!("Cannot divide current value by zero");
                 } else if divisor < 0 {
@@ -255,7 +255,7 @@ macro_rules! impl_div_for_integer {
                 }
                 self.raw
                     .checked_div(divisor as u64)
-                    .map(Current::from_micro_amps)
+                    .map(Self::from_micro_amps)
                     .expect("Overflow when dividing current value")
             }
         }
@@ -272,21 +272,21 @@ impl_div_for_integer!(i32);
 impl_div_for_integer!(i64);
 
 impl ops::Div<f32> for Current {
-    type Output = Current;
+    type Output = Self;
 
     /// Divides a `Current` value by a floating-point value, returning a new `Current` value.
     #[inline]
-    fn div(self, divisor: f32) -> Current {
+    fn div(self, divisor: f32) -> Self {
         self / divisor as f64
     }
 }
 
 impl ops::Div<f64> for Current {
-    type Output = Current;
+    type Output = Self;
 
     /// Divides a `Current` value by a floating-point value, returning a new `Current` value.
     #[inline]
-    fn div(self, divisor: f64) -> Current {
+    fn div(self, divisor: f64) -> Self {
         let result = match divisor {
             _ if divisor == 0f64 => panic!("Cannot divide current value by zero"),
             _ if divisor.is_infinite() => {
@@ -299,11 +299,11 @@ impl ops::Div<f64> for Current {
             _ => (self.raw as f64) / divisor,
         };
 
-        Current::from_micro_amps(result as u64)
+        Self::from_micro_amps(result as u64)
     }
 }
 
-/// Extension trait for simple short-hands for creating `Current` values from `u32` values.
+/// Extension trait for simple short-hands for creating `Current` values from integer values.
 pub trait FromInteger {
     /// Creates a new `Current` from a number of whole microamps (μA).
     fn micro_amps(self) -> Current;
@@ -351,7 +351,7 @@ impl_current_from_integer!(i16);
 impl_current_from_integer!(i32);
 impl_current_from_integer!(i64);
 
-/// Extension trait for simple short-hands for creating `Current` values from `f32` values.
+/// Extension trait for simple short-hands for creating `Current` values from floating-point values.
 pub trait FromFloat {
     /// Creates a new `Current` from a number of fractional microamps (μA).
     ///