+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this library will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - 2023-02-23
+
+### Added
+
+- `Power` struct for storing power values in microwatts (μW)
+- `Power` extension methods for integer and floating point types via macros
+- Ohm's Law calculations for `Power` via operators (`Voltage * Current` and vice versa)
+
+### Changed
+
+- Some documentation improvements
+
 ## [0.3.2] - 2023-02-23
 
 ### Fixed